Added a test for the meta stuff and fixed a couple bugs.  Somehow it works which seems silly.  Feel like gotta be missing something

+++ netty/src/main/java/com/urbanairship/hbase/shc/dispatch/netty/pipeline/HbaseResponseDecoder.java
@@ -89,7 +89,7 @@ public final class HbaseResponseDecoder extends OneToOneDecoder {
     }
 
     private String readFramedString(ChannelBuffer buffer) {
-        int length = buffer.readShort();
+        int length = buffer.readInt();
         Preconditions.checkArgument(length > 0);
 
         byte[] bytes = new byte[length];
